example of use hook - review
